+++ UC4-chaincode/src/main/java/de/upb/cs/uc4/chaincode/model/Approval.java
@@ -3,58 +3,48 @@ package de.upb.cs.uc4.chaincode.model;
 import com.google.gson.annotations.SerializedName;
 import io.swagger.annotations.ApiModelProperty;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 
 /**
- * DetailedError
+ * MatriculationData
  */
 @javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaClientCodegen", date = "2020-07-26T19:00:46.792+02:00")
 
 
 
-public class GenericError {
+public class Approval {
+  @SerializedName("id")
+  private String id = null;
+
+  public Approval id(String id) {
+    this.id = id;
+    return this;
+  }
+
+  @ApiModelProperty(value = "")
+  public String getId() {
+    return id;
+  }
+
+  public void setId(String id) {
+    this.id = id;
+  }
+
   @SerializedName("type")
   private String type = null;
 
-  @SerializedName("title")
-  private String title = null;
-
-  public GenericError type(String type) {
+  public Approval type(String type) {
     this.type = type;
     return this;
   }
 
-   /**
-   * Get type
-   * @return type
-  **/
   @ApiModelProperty(value = "")
   public String getType() {
     return type;
   }
 
-  public void setType(String type) {
+  public void setIType(String type) {
     this.type = type;
-  }
-
-  public GenericError title(String title) {
-    this.title = title;
-    return this;
-  }
-
-   /**
-   * Get title
-   * @return title
-  **/
-  @ApiModelProperty(value = "")
-  public String getTitle() {
-    return title;
-  }
-
-  public void setTitle(String title) {
-    this.title = title;
   }
 
 
@@ -66,24 +56,24 @@ public class GenericError {
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    GenericError detailedError = (GenericError) o;
-    return Objects.equals(this.type, detailedError.type) &&
-        Objects.equals(this.title, detailedError.title);
+    Approval matriculationData = (Approval) o;
+    return Objects.equals(this.id, matriculationData.id) &&
+        Objects.equals(this.type, matriculationData.type);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(type, title);
+    return Objects.hash(id, type);
   }
 
 
   @Override
   public String toString() {
     StringBuilder sb = new StringBuilder();
-    sb.append("class DetailedError {\n");
+    sb.append("class MatriculationData {\n");
     
+    sb.append("    id: ").append(toIndentedString(id)).append("\n");
     sb.append("    type: ").append(toIndentedString(type)).append("\n");
-    sb.append("    title: ").append(toIndentedString(title)).append("\n");
     sb.append("}");
     return sb.toString();
   }